+++ Projeto/02-Final-version/gerenpilas/app/Http/Controllers/TransactionController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Transaction;
 use Illuminate\Http\Request;
 use App\Models\Category;
+use Illuminate\Support\Facades\Auth;
 
 class TransactionController extends Controller
 {
@@ -15,8 +16,34 @@ class TransactionController extends Controller
      */
     public function index()
     {
-        $transactions = Transaction::orderBy('data')->get();
-        return view('principal', ['transactions'=>$transactions]);
+        if (Auth::check()) {
+            $usuario_logado = Auth::user()->id;
+            $data = array();
+
+            if (empty($_GET)) {
+                $data['mes'] = date('m');
+                $data['ano'] = date('Y');
+            } else {
+                try{
+                    $data['mes'] = $_GET['mes'];
+                    $data['ano'] = $_GET['ano'];
+                } catch (\Exception $e) {
+                    $data['mes'] = date('m');
+                    $data['ano'] = date('Y');
+                }
+            }
+
+            $transactions = Transaction::where('user_id', $usuario_logado)->whereMonth('data', $data['mes'])->whereYear('data', $data['ano'])->orderBy('data')->get();
+            $categories = Category::where('user_id', $usuario_logado)->orderBy('nome')->get();
+            $despesa_mensal = Transaction::where('user_id', $usuario_logado)->whereMonth('data', $data['mes'])->whereYear('data', $data['ano'])->whereTipo(0)->sum('valor');
+            $receita_mensal = Transaction::where('user_id', $usuario_logado)->whereMonth('data', $data['mes'])->whereYear('data', $data['ano'])->whereTipo(1)->sum('valor');
+            $despesa_total = Transaction::where('user_id', $usuario_logado)->whereTipo(0)->sum('valor');
+            $receita_total = Transaction::where('user_id', $usuario_logado)->whereTipo(1)->sum('valor');
+
+            return view('home', compact('transactions', 'categories', 'despesa_mensal', 'receita_mensal', 'despesa_total', 'receita_total', 'data'));
+        } else {
+            return view('auth.login');
+        }
     }
 
     /**
@@ -26,8 +53,12 @@ class TransactionController extends Controller
      */
     public function create()
     {
-        $categorias = Category::orderBy('nome')->get();
-        return view('transactions.create', ['categorias'=>$categorias]);
+        if (Auth::check()) {
+            $categorias = Category::where('user_id', Auth::user()->id)->orderBy('nome')->get();
+            return view('transactions.create', ['categorias'=>$categorias]);
+        } else {
+            return view('auth.login');
+        }
     }
 
     /**
@@ -38,16 +69,20 @@ class TransactionController extends Controller
      */
     public function store(Request $request)
     {
-        $transaction = new Transaction;
-        $transaction->user_id = $request->user_id;
-        $transaction->category_id = $request->category_id;
-        $transaction->data = $request->data;
-        $transaction->descricao = $request->descricao;
-        $transaction->valor = $request->valor;
-        $transaction->tipo = intval($request->tipo);
-        $transaction->save();
-        session()->flash('msg-success', 'Transação gravada com sucesso!');
-        return redirect()->route('index');
+        if (Auth::check()) {
+            $transaction = new Transaction;
+            $transaction->user_id = Auth::user()->id;
+            $transaction->category_id = $request->category_id;
+            $transaction->data = $request->data;
+            $transaction->descricao = $request->descricao;
+            $transaction->valor = number_format(str_replace(',', '.', $request->valor), 2, ".", "");
+            $transaction->tipo = intval($request->tipo);
+            $transaction->save();
+            session()->flash('msg-success', 'Transação gravada com sucesso!');
+            return redirect()->route('index');
+        } else {
+            return view('auth.login');
+        }
     }
 
     /**
@@ -58,7 +93,7 @@ class TransactionController extends Controller
      */
     public function show(Transaction $transaction)
     {
-        //
+        return redirect()->route('index');
     }
 
     /**
@@ -69,8 +104,12 @@ class TransactionController extends Controller
      */
     public function edit(Transaction $transaction)
     {
-        $categorias = Category::orderBy('nome')->get();
-        return view('transactions.edit', ['transaction'=>$transaction, 'categorias'=>$categorias]);
+        if (Auth::check()) {
+            $categorias = Category::where('user_id', Auth::user()->id)->orderBy('nome')->get();
+            return view('transactions.edit', ['transaction'=>$transaction, 'categorias'=>$categorias]);
+        } else {
+            return view('auth.login');
+        }
     }
 
     /**
@@ -82,14 +121,19 @@ class TransactionController extends Controller
      */
     public function update(Request $request, Transaction $transaction)
     {
-        $transaction->category_id = $request->category_id;
-        $transaction->data = $request->data;
-        $transaction->descricao = $request->descricao;
-        $transaction->valor = $request->valor;
-        $transaction->tipo = intval($request->tipo);
-        $transaction->save();
-        session()->flash('msg-success', 'Suas alterações foram gravadas com sucesso!');
-        return redirect()->route('index');
+        if (Auth::check()) {
+            $transaction->data = $request->data;
+            // $transaction->category_id = Auth::user()->id;
+            $transaction->descricao = $request->descricao;
+            $transaction->valor = number_format(str_replace(',', '.', $request->valor), 2, ".", "");
+            $transaction->category_id = $request->category_id;
+            $transaction->tipo = intval($request->tipo);
+            $transaction->save();
+            session()->flash('msg-success', 'Suas alterações foram gravadas com sucesso!');
+            return redirect()->route('index');
+        } else {
+            return view('auth.login');
+        }
     }
 
     /**
@@ -100,8 +144,12 @@ class TransactionController extends Controller
      */
     public function destroy(Transaction $transaction)
     {
-        $transaction->delete();
-        session()->flash('msg-success', 'Transação excluída!');
-        return redirect()->route('index');
+        if (Auth::check()) {
+            $transaction->delete();
+            session()->flash('msg-success', 'Transação excluída!');
+            return redirect()->route('index');
+        } else {
+            return view('auth.login');
+        }
     }
 }
